Add more unit tests and logs
